refactor(web): extract AgentsTree + AgentDetailPane components
Pure refactor with byte-identical render. Two new component files
break the agents tree rendering and the agent detail rendering out
of their page wrappers so they can be composed into a split-view
parent in a later commit:

- components/AgentsTree.tsx — pure tree rendering. Exports
  `buildTree`, `relativeTime`, and a default `AgentsTree` that
  takes a `tree: TreeNode[]` prop. RunList now feeds in the result
  of `buildTree(agents)`.

- components/AgentDetailPane.tsx — full agent detail body (header,
  events, cancel button, polling, coalesceText, EventCard). Takes
  `agentId` as a prop instead of reading useParams.

Page wrappers shrink to thin shells:
- pages/RunList.tsx keeps the filter state + polling effect, renders
  the toolbar + <AgentsTree>.
- pages/AgentDetail.tsx becomes a 12-line shell that reads the URL
  param and renders <AgentDetailPane agentId={agentId}/>.

No behaviour change — every render output is identical to v0.8.19.
The Link target on tree nodes still points to /agents/:id; the
detail pane still polls at 1.5s; the breadcrumb is still the
single "← runs" link from before. Per-node expand/collapse +
breadcrumbs + split layout come in subsequent commits.

`cd web && npm run build` clean (50 modules, no warnings).
